% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_10_demo_source.R
\name{testFunction}
\alias{testFunction}
\title{Source code for module 10 demo package}
\usage{
testFunction(x)
}
\arguments{
\item{x}{numerical value}
}
\value{
returns the given x value multiplied by itself
}
\description{
Source code for module 10 demo package
}
\examples{
testFunction(5)
testFunction(9)
}
